feat(adherent): remove old and add new page in blog (temporaly)
